Rework logic related to ImapFolderEvent entity Haulmont/jmix-imap#1
1. Add the "enabled" flag for the ImapFolderEvent entity. This flag specifies whether this type of event should be generated for a folder or not.
2. Rework ImapMailBox editor: Folders table containing generated columns for each event type is replaced by three GroupBoxes: Folders, Events and Handlers. Each GroupBox contains the table.
the Events table is updated when the folder in the Folders table is selected. The Handlers table is updated when the event in the Events table is selected.
There is an ability to enable/disable all events for a folder by select the required action in the actions list in the Events table.
3. Add a test case in ImapEventsTest when the "enabled" flag for ImapFolderEvent is set to false.